Change dkim_libversion() from u_long to uint32_t, and document how the version is encoded.

+++ libopendkim/dkim.c
@@ -8424,7 +8424,7 @@ dkim_libfeature(DKIM_LIB *lib, u_int fc)
 **  	Library version, i.e. value of the OPENDKIM_LIB_VERSION macro.
 */
 
-unsigned long
+uint32_t
 dkim_libversion(void)
 {
 	return OPENDKIM_LIB_VERSION;
